changed encoders to return int for pulses, safeguard is coded

+++ 2018 Competition Robot/src/org/usfirst/frc/team1732/robot/sensors/encoders/EncoderBase.java
@@ -19,7 +19,7 @@ public abstract class EncoderBase {
 	 */
 	public abstract double getRate();
 
-	public abstract double getPulses();
+	public abstract int getPulses();
 
 	/**
 	 * @param distancePerPulse
